Refactor the resolveJsonLinks method
The resolveJsonLinks() only handled (Link.types). This is problematic
because any experiment with new annotation is therefore impossible.

Any annotation thas is matching the regexp will now be dropped:

 \((?:([a-zA-Z0-9]+)\.)?([a-zA-Z0-9]+)\)

I took the time to refactor the method as well to avoid using forEach
and "holder" arrays or atomic references. It's a bit more verbose but
easier to read IMO.